+++ src/main/kotlin/com/terraformation/backend/tracking/api/BiomassPayloads.kt
@@ -1,0 +1,320 @@
+package com.terraformation.backend.tracking.api
+
+import com.fasterxml.jackson.annotation.JsonTypeName
+import com.terraformation.backend.db.default_schema.SpeciesId
+import com.terraformation.backend.db.tracking.BiomassForestType
+import com.terraformation.backend.db.tracking.MangroveTide
+import com.terraformation.backend.db.tracking.ObservationPlotPosition
+import com.terraformation.backend.db.tracking.TreeGrowthForm
+import com.terraformation.backend.tracking.model.BiomassQuadratModel
+import com.terraformation.backend.tracking.model.BiomassQuadratSpeciesModel
+import com.terraformation.backend.tracking.model.BiomassSpeciesModel
+import com.terraformation.backend.tracking.model.ExistingBiomassDetailsModel
+import com.terraformation.backend.tracking.model.ExistingRecordedTreeModel
+import com.terraformation.backend.tracking.model.NewBiomassDetailsModel
+import com.terraformation.backend.tracking.model.NewRecordedTreeModel
+import io.swagger.v3.oas.annotations.media.Schema
+import java.math.BigDecimal
+import java.time.Instant
+
+data class BiomassQuadratPayload(
+    val description: String?,
+    val position: ObservationPlotPosition,
+    val species: List<BiomassQuadratSpeciesPayload>,
+) {
+  constructor(
+      model: BiomassQuadratModel,
+      position: ObservationPlotPosition
+  ) : this(
+      description = model.description,
+      position = position,
+      species = model.species.map { BiomassQuadratSpeciesPayload(it) },
+  )
+
+  fun toModel(): BiomassQuadratModel {
+    return BiomassQuadratModel(
+        description = description,
+        species = species.map { it.toModel() }.toSet(),
+    )
+  }
+}
+
+data class BiomassQuadratSpeciesPayload(
+    @Schema(minimum = "0", maximum = "100") //
+    val abundancePercent: Int,
+    val speciesId: SpeciesId?,
+    val speciesName: String?,
+) {
+  constructor(
+      model: BiomassQuadratSpeciesModel,
+  ) : this(
+      abundancePercent = model.abundancePercent,
+      speciesId = model.speciesId,
+      speciesName = model.speciesName,
+  )
+
+  fun toModel(): BiomassQuadratSpeciesModel {
+    return BiomassQuadratSpeciesModel(
+        abundancePercent = abundancePercent,
+        speciesId = speciesId,
+        speciesName = speciesName,
+    )
+  }
+}
+
+data class BiomassSpeciesPayload(
+    val commonName: String?,
+    val isInvasive: Boolean,
+    val isThreatened: Boolean,
+    val scientificName: String?,
+    val speciesId: SpeciesId?,
+) {
+  constructor(
+      model: BiomassSpeciesModel
+  ) : this(
+      isInvasive = model.isInvasive,
+      isThreatened = model.isThreatened,
+      speciesId = model.speciesId,
+      scientificName = model.scientificName,
+      commonName = model.commonName,
+  )
+
+  fun toModel(): BiomassSpeciesModel {
+    return BiomassSpeciesModel(
+        commonName = commonName,
+        isInvasive = isInvasive,
+        isThreatened = isThreatened,
+        scientificName = scientificName,
+        speciesId = speciesId,
+    )
+  }
+}
+
+data class ExistingTreePayload(
+    val description: String?,
+    @Schema(description = "Measured in centimeters, required if growth form is Tree or Trunk.")
+    val diameterAtBreastHeight: BigDecimal?,
+    @Schema(description = "Measured in meters, required if growth form is Tree.")
+    val height: BigDecimal?,
+    val isDead: Boolean,
+    @Schema(description = "Measured in meters, required if growth form is Tree or Trunk.")
+    val pointOfMeasurement: BigDecimal?,
+    @Schema(description = "Measured in centimeters, required if growth form is Shrub.")
+    val shrubDiameter: Int?,
+    val speciesId: SpeciesId?,
+    val speciesName: String?,
+    val treeGrowthForm: TreeGrowthForm,
+    @Schema(description = "Tree identifier number. Must be unique if growth form is not Trunk.")
+    val treeNumber: Int,
+    @Schema(
+        description =
+            "Trunk identifier number. Must be unique. Must be 1 if growth form " + "is not Trunk.")
+    val trunkNumber: Int
+) {
+  constructor(
+      model: ExistingRecordedTreeModel
+  ) : this(
+      description = model.description,
+      diameterAtBreastHeight = model.diameterAtBreastHeightCm,
+      height = model.heightM,
+      isDead = model.isDead,
+      pointOfMeasurement = model.pointOfMeasurementM,
+      shrubDiameter = model.shrubDiameterCm,
+      speciesId = model.speciesId,
+      speciesName = model.speciesName,
+      treeGrowthForm = model.treeGrowthForm,
+      treeNumber = model.treeNumber,
+      trunkNumber = model.trunkNumber,
+  )
+}
+
+interface NewTreePayload {
+  val description: String?
+  val isDead: Boolean?
+  val speciesId: SpeciesId?
+  val speciesName: String?
+
+  fun toTreeModels(treeNumber: Int): List<NewRecordedTreeModel>
+}
+
+@JsonTypeName("shrub")
+data class NewShrubPayload(
+    override val description: String?,
+    override val isDead: Boolean,
+    override val speciesId: SpeciesId?,
+    override val speciesName: String?,
+    @Schema(description = "Measured in centimeters.") val shrubDiameter: Int,
+) : NewTreePayload {
+  override fun toTreeModels(treeNumber: Int): List<NewRecordedTreeModel> {
+    return listOf(
+        NewRecordedTreeModel(
+            id = null,
+            description = description,
+            diameterAtBreastHeightCm = null,
+            heightM = null,
+            isDead = isDead,
+            pointOfMeasurementM = null,
+            shrubDiameterCm = shrubDiameter,
+            speciesId = speciesId,
+            speciesName = speciesName,
+            treeGrowthForm = TreeGrowthForm.Shrub,
+            treeNumber = treeNumber,
+            trunkNumber = 1,
+        ))
+  }
+}
+
+@JsonTypeName("singleTrunk")
+data class NewSingleTrunkTreePayload(
+    override val description: String?,
+    @Schema(description = "Measured in centimeters.") val diameterAtBreastHeight: BigDecimal,
+    override val isDead: Boolean,
+    override val speciesId: SpeciesId?,
+    override val speciesName: String?,
+    @Schema(description = "Measured in meters.") val height: BigDecimal,
+    @Schema(description = "Measured in meters.") val pointOfMeasurement: BigDecimal,
+) : NewTreePayload {
+  override fun toTreeModels(treeNumber: Int): List<NewRecordedTreeModel> {
+    return listOf(
+        NewRecordedTreeModel(
+            id = null,
+            description = description,
+            diameterAtBreastHeightCm = diameterAtBreastHeight,
+            heightM = height,
+            isDead = isDead,
+            pointOfMeasurementM = pointOfMeasurement,
+            shrubDiameterCm = null,
+            speciesId = speciesId,
+            speciesName = speciesName,
+            treeGrowthForm = TreeGrowthForm.Shrub,
+            treeNumber = treeNumber,
+            trunkNumber = 1,
+        ))
+  }
+}
+
+data class NewTrunkPayload(
+    @Schema(description = "Measured in centimeters.") val diameterAtBreastHeight: BigDecimal,
+    @Schema(description = "Measured in meters.") val pointOfMeasurement: BigDecimal,
+) {
+  fun toTreeModel(
+      parent: NewMultiTrunksTreePayload,
+      treeNumber: Int,
+      trunkNumber: Int
+  ): NewRecordedTreeModel {
+    return NewRecordedTreeModel(
+        id = null,
+        description = parent.description,
+        diameterAtBreastHeightCm = diameterAtBreastHeight,
+        heightM = null,
+        isDead = parent.isDead,
+        pointOfMeasurementM = pointOfMeasurement,
+        shrubDiameterCm = null,
+        speciesId = parent.speciesId,
+        speciesName = parent.speciesName,
+        treeGrowthForm = TreeGrowthForm.Shrub,
+        treeNumber = treeNumber,
+        trunkNumber = trunkNumber,
+    )
+  }
+}
+
+@JsonTypeName("multiTrunks")
+data class NewMultiTrunksTreePayload(
+    override val description: String?,
+    override val isDead: Boolean,
+    override val speciesId: SpeciesId?,
+    override val speciesName: String?,
+    val trunks: List<NewTrunkPayload>,
+) : NewTreePayload {
+  override fun toTreeModels(treeNumber: Int): List<NewRecordedTreeModel> {
+    return trunks.mapIndexed { index, trunk -> trunk.toTreeModel(this, treeNumber, index + 1) }
+  }
+}
+
+data class ExistingBiomassMeasurementPayload(
+    val description: String?,
+    val forestType: BiomassForestType,
+    val herbaceousCoverPercent: Int,
+    val ph: BigDecimal?,
+    val quadrats: List<BiomassQuadratPayload>,
+    @Schema(description = "Measured in ppt") val salinity: BigDecimal?,
+    val smallTreeCountLow: Int,
+    val smallTreeCountHigh: Int,
+    val soilAssessment: String,
+    @Schema(description = "Low or high tide.") val tide: MangroveTide?,
+    @Schema(description = "Time when ide is observed.") val tideTime: Instant?,
+    val trees: List<ExistingTreePayload>,
+    val species: List<BiomassSpeciesPayload>,
+    @Schema(description = "Measured in centimeters.") val waterDepth: Int?,
+) {
+  constructor(
+      model: ExistingBiomassDetailsModel
+  ) : this(
+      description = model.description,
+      forestType = model.forestType,
+      herbaceousCoverPercent = model.herbaceousCoverPercent,
+      ph = model.ph,
+      quadrats =
+          model.quadrats.map { (position, quadrat) -> BiomassQuadratPayload(quadrat, position) },
+      salinity = model.salinityPpt,
+      smallTreeCountLow = model.smallTreeCountRange.first,
+      smallTreeCountHigh = model.smallTreeCountRange.second,
+      soilAssessment = model.soilAssessment,
+      species = model.species.map { BiomassSpeciesPayload(it) },
+      tide = model.tide,
+      tideTime = model.tideTime,
+      trees = model.trees.map { ExistingTreePayload(it) },
+      waterDepth = model.waterDepthCm,
+  )
+}
+
+data class NewBiomassMeasurementPayload(
+    val description: String?,
+    val forestType: BiomassForestType,
+    @Schema(minimum = "0", maximum = "100") //
+    val herbaceousCoverPercent: Int,
+    @Schema(description = "Required for Mangrove forest.", minimum = "0", maximum = "14")
+    val ph: BigDecimal?,
+    val quadrats: List<BiomassQuadratPayload>,
+    @Schema(description = "Measured in ppt. Required for Mangrove forest.", minimum = "0")
+    val salinity: BigDecimal?,
+    @Schema(minimum = "0") //
+    val smallTreeCountLow: Int,
+    @Schema(minimum = "smallTreeCountHigh") //
+    val smallTreeCountHigh: Int,
+    val soilAssessment: String,
+    @Schema(description = "Low or high tide. Required for Mangrove forest.")
+    val tide: MangroveTide?,
+    @Schema(description = "Time when ide is observed. Required for Mangrove forest.")
+    val tideTime: Instant?,
+    val trees: List<NewTreePayload>,
+    @Schema(
+        description =
+            "List of herbaceous and tree species. Includes all recorded quadrat and additional " +
+                "herbaceous species and recorded tree species. Species not assigned to a quadrat or " +
+                "recorded trees will be saved as an additional herbaceous species.")
+    val species: List<BiomassSpeciesPayload>,
+    @Schema(description = "Measured in centimeters. Required for Mangrove forest.")
+    val waterDepth: Int?,
+) {
+  fun toModel(): NewBiomassDetailsModel {
+    return NewBiomassDetailsModel(
+        description = description,
+        forestType = forestType,
+        herbaceousCoverPercent = herbaceousCoverPercent,
+        observationId = null,
+        ph = ph,
+        quadrats = quadrats.associateBy { it.position }.mapValues { it.value.toModel() },
+        salinityPpt = salinity,
+        smallTreeCountRange = smallTreeCountLow to smallTreeCountHigh,
+        soilAssessment = soilAssessment,
+        species = species.map { it.toModel() }.toSet(),
+        plotId = null,
+        tide = tide,
+        tideTime = tideTime,
+        trees = trees.flatMapIndexed { index, tree -> tree.toTreeModels(index + 1) },
+        waterDepthCm = waterDepth,
+    )
+  }
+}
